pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "corp-knowledge-extractor"
version = "1.0.0"
description = "Corporate knowledge extraction engine — tiered Gemini pipeline"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
    "google-genai>=1.0.0",
    "click>=8.0",
    "rich>=13.0",
    "Jinja2>=3.0",
    "opencv-python>=4.8.0",
    "pydub>=0.25.1",
    "numpy>=1.24",
    "requests>=2.28",
    "pillow>=9.0",
    "packaging>=23.0",
    "pdfplumber>=0.10.0",
    "python-pptx>=1.0.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff", "mypy"]
slides = ["comtypes; sys_platform == 'win32'", "PyMuPDF>=1.20.0"]

[project.scripts]
cke = "scripts.run:cli"

[tool.setuptools.packages.find]
include = ["src*", "scripts*", "config*", "templates*"]

[tool.ruff]
line-length = 120

[tool.mypy]
python_version = "3.11"
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]
